:art: :: 모듈 개편 작업
- 모듈 수직 분할(data-domain)
- 폴더링
