+++ build-logic/src/main/kotlin/convention.android.feature.gradle.kts
@@ -11,10 +11,10 @@ android {
 }
 
 dependencies {
+    "implementation"(project(":data"))
+    "implementation"(project(":domain"))
     "implementation"(project(":common:resource"))
     "implementation"(project(":core:designsystem"))
-    "implementation"(project(":core:data"))
-    "implementation"(project(":core:domain"))
     "implementation"(project(":core:util"))
     implementation(project(":core:navigation"))
 }
