--- core/domain/build.gradle.kts
@@ -1,15 +1,0 @@
-plugins {
-    id("convention.android.library")
-}
-
-android {
-    namespace = "com.idiotfrogs.domain"
-}
-
-dependencies {
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.test.ext)
-    androidTestImplementation(libs.androidx.test.espresso)
-}
