--- core/domain/src/main/java/com/idiotfrogs/domain/exception/LoginCancelledException.kt
@@ -1,3 +1,0 @@
-package com.idiotfrogs.domain.exception
-
-class LoginCancelledException: Exception("User cancelled login")
