+++ data/src/main/java/com/idiotfrogs/data/exception/LoginCancelledException.kt
@@ -1,0 +1,3 @@
+package com.idiotfrogs.data.exception
+
+class LoginCancelledException: Exception("User cancelled login")
