+++ domain/src/androidTest/java/com/idiotfrogs/domain/ExampleInstrumentedTest.kt
