+++ domain/src/main/AndroidManifest.xml
