+++ feature/auth/src/main/java/com/idiotfrogs/auth/util/RememberLoginManager.kt
@@ -3,7 +3,7 @@ package com.idiotfrogs.auth.util
 import androidx.activity.compose.LocalActivity
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import com.idiotfrogs.data.LoginManager
+import com.idiotfrogs.data.remote.LoginManager
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.android.EntryPointAccessors
